pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "assetsmanager"
version = "0.1.0"
description = "Professional asset management for video editors"
requires-python = ">=3.11"
dependencies = [
    "PySide6>=6.6.0",
    "SQLAlchemy>=2.0.0",
    "Pillow>=10.0.0",
    "av>=11.0.0",
]

[project.scripts]
assetsmanager = "main:main"

[tool.setuptools.packages.find]
where = ["."]
